duplicates module
