+++ docs/reference/geopandas_tools/duplicates.rst
@@ -1,0 +1,7 @@
+Handle duplicate and overlapping geometries
+===========================================
+
+.. automodule:: sgis.geopandas_tools.duplicates
+    :members:
+    :exclude-members:
+    :show-inheritance:
