--- docs/reference/multiprocessing/mapper.rst
@@ -1,7 +1,0 @@
-MultiProcessingMapper
-=====================
-
-.. automodule:: sgis.multiprocessing.multiprocessingmapper
-    :members:
-    :exclude-members:
-    :show-inheritance:
