+++ docs/reference/parallel/index.rst
@@ -1,8 +1,7 @@
-Multiprocessing
+Parallelization
 ===============
 
 .. toctree::
    :maxdepth: 3
 
-   pool
-   mapper
+   parallel
