+++ src/sgis/geopandas_tools/overlay.py
@@ -7,223 +7,16 @@ version of the solution from GH 2792.
 'clean_overlay' also includes the overlay type "update", which can be specified in the
 "how" parameter, in addition to the five native geopandas how-s.
 """
-from typing import Callable
-
 import geopandas as gpd
 import numpy as np
 import pandas as pd
 from geopandas import GeoDataFrame
 from pandas import DataFrame
 from pyproj import CRS
-from shapely import STRtree, difference, intersection, make_valid, unary_union, union
-from shapely.errors import GEOSException
-from shapely.geometry import Polygon
+from shapely import STRtree, difference, intersection, make_valid, unary_union
 
 from .general import clean_geoms
 from .geometry_types import get_geom_type, make_all_singlepart, to_single_geom_type
-
-
-def update_geometries(
-    gdf: GeoDataFrame, keep_geom_type: bool = True, copy: bool = True
-) -> GeoDataFrame:
-    """Puts geometries on top of each other rowwise.
-
-    Since this operation is done rowwise, it's important to
-    first sort the GeoDataFrame approriately. See example below.
-
-    Example
-    ------
-    Create two circles and get the overlap.
-
-    >>> import sgis as sg
-    >>> circles = sg.to_gdf([(0, 0), (1, 1)]).pipe(sg.buff, 1)
-    >>> duplicates = sg.get_intersections(circles)
-    >>> duplicates
-       idx                                           geometry
-    0    1  POLYGON ((0.03141 0.99951, 0.06279 0.99803, 0....
-    1    2  POLYGON ((1.00000 0.00000, 0.96859 0.00049, 0....
-
-    The polygons are identical except for the order of the coordinates.
-
-    >>> poly1, poly2 = duplicates.geometry
-    >>> poly1.equals(poly2)
-    True
-
-    'update_geometries' gives different results based on the order
-    of the GeoDataFrame.
-
-    >>> sg.update_geometries(duplicates)
-        idx                                           geometry
-    0    1  POLYGON ((0.03141 0.99951, 0.06279 0.99803, 0....
-
-    >>> dups_rev = duplicates.iloc[::-1]
-    >>> sg.update_geometries(dups_rev)
-        idx                                           geometry
-    1    2  POLYGON ((1.00000 0.00000, 0.96859 0.00049, 0....
-
-    It might be appropriate to put the largest polygons on top
-    and sort all NaNs to the bottom.
-
-    >>> updated = (
-    ...     sg.sort_large_first(duplicates)
-    ...     .pipe(sg.sort_nans_last)
-    ...     .pipe(sg.update_geometries)
-    >>> updated
-        idx                                           geometry
-    0    1  POLYGON ((0.03141 0.99951, 0.06279 0.99803, 0....
-
-    """
-    if len(gdf) <= 1:
-        return gdf
-
-    df = pd.DataFrame(gdf, copy=copy)
-
-    unioned = Polygon()
-    out_rows, indices, geometries = [], [], []
-
-    if keep_geom_type:
-        geom_type = get_geom_type(gdf)
-        if geom_type == "mixed":
-            raise ValueError("Cannot have mixed geometries when keep_geom_type is True")
-
-    for i, row in df.iterrows():
-        geom = row.pop("geometry")
-
-        if any(geom.equals(geom2) for geom2 in geometries):
-            continue
-
-        new = try_shapely_func_pair(geom, unioned, func=difference)
-
-        if not new:
-            continue
-
-        unioned = try_shapely_func_pair(new, unioned, func=union)
-
-        out_rows.append(row)
-        geometries.append(new)
-        indices.append(i)
-
-    out = GeoDataFrame(out_rows, geometry=geometries, index=indices, crs=gdf.crs)
-
-    if keep_geom_type:
-        out = to_single_geom_type(out, geom_type)
-
-    return out
-
-
-def try_shapely_func_pair(geom1, geom2, func: Callable):
-    try:
-        return func(geom1, geom2)
-    except GEOSException:
-        try:
-            geom1 = make_valid(geom1)
-            return func(geom1, geom2)
-        except GEOSException:
-            try:
-                return func(geom1, geom2, grid_size=0.01)
-            except GEOSException as e:
-                raise ValueError(geom1, geom2) from e
-
-
-def get_intersections(gdf: GeoDataFrame) -> GeoDataFrame:
-    """Find geometries that intersect in a GeoDataFrame.
-
-    Does an intersection with itself and keeps only the geometries that appear
-    more than once. This means each intersection gives at least two rows. The
-    duplicates should then be handled appropriately. See example
-    below.
-
-    Args:
-        gdf: GeoDataFrame of polygons.
-
-    Returns:
-        A GeoDataFrame of the overlapping polygons.
-
-    Examples
-    --------
-    Create two fully overlapping polygons.
-
-    >>> import sgis as sg
-    >>> circles = sg.to_gdf([(0, 0), (0, 0)])
-    >>> circles["geometry"] = circles["geometry"].buffer([1, 2])
-    >>> circles["idx"] = [1, 3]
-    >>> circles.area
-    0     3.141076
-    1    12.564304
-    dtype: float64
-
-    Get the duplicates.
-
-    >>> duplicates = sg.get_intersections(circles)
-    >>> duplicates
-       idx                                           geometry
-    0    1  POLYGON ((0.99951 -0.03141, 0.99803 -0.06279, ...
-    1    3  POLYGON ((0.99951 -0.03141, 0.99803 -0.06279, ...
-    >>> duplicates.area
-    0    3.141076
-    1    3.141076
-    dtype: float64
-
-    We get two rows for each intersection pair.
-
-    The function sgis.update_geometries can be used to put geometries
-    on top of the other rowwise.
-
-    >>> updated = sg.update_geometries(duplicates)
-    >>> updated
-        idx                                           geometry
-    0    1  POLYGON ((0.99518 -0.09802, 0.98079 -0.19509, ...
-
-    Reversing the rows means the bottom polygon is put on top.
-
-    >>> updated = sg.update_geometries(duplicates.iloc[::-1]))
-    >>> updated
-        idx                                           geometry
-    1    3  POLYGON ((0.99518 -0.09802, 0.98079 -0.19509, ...
-
-    It might be appropriate to sort the dataframe by columns.
-    Or put large polygons first and NaN values last.
-
-    >>> updated = (
-    ...     sg.sort_large_first(duplicates)
-    ...     .pipe(sg.sort_nans_last)
-    ...     .pipe(sg.update_geometries)
-    ... )
-    >>> updated
-       idx                                           geometry
-    0    1  POLYGON ((0.99518 -0.09802, 0.98079 -0.19509, ...
-    """
-
-    idx_name = gdf.index.name
-    duplicated_geoms = _get_intersecting_geometries(gdf).pipe(clean_geoms)
-
-    duplicated_geoms.index = duplicated_geoms["orig_idx"].values
-    duplicated_geoms.index.name = idx_name
-    return duplicated_geoms.drop(columns="orig_idx")
-
-
-def _get_intersecting_geometries(gdf: GeoDataFrame) -> GeoDataFrame:
-    gdf = gdf.assign(orig_idx=gdf.index).reset_index(drop=True)
-
-    right = gdf[[gdf._geometry_column_name]]
-    right["idx_right"] = right.index
-    left = gdf
-    left["idx_left"] = left.index
-
-    intersected = clean_overlay(left, right, how="intersection")
-
-    # these are identical as the input geometries
-    not_from_same_poly = intersected.loc[lambda x: x["idx_left"] != x["idx_right"]]
-
-    points_joined = (
-        not_from_same_poly.representative_point().to_frame().sjoin(not_from_same_poly)
-    )
-
-    duplicated_points = points_joined.loc[points_joined.index.duplicated(keep=False)]
-
-    return intersected.loc[intersected.index.isin(duplicated_points.index)].drop(
-        columns=["idx_left", "idx_right"]
-    )
 
 
 def clean_overlay(
