+++ src/sgis/geopandas_tools/polygon_operations.py
@@ -20,7 +20,7 @@ def get_polygon_clusters(
     *gdfs: GeoDataFrame | GeoSeries,
     cluster_col: str = "cluster",
     allow_multipart: bool = False,
-    wkt_col: bool = False,
+    as_string: bool = False,
 ) -> GeoDataFrame | tuple[GeoDataFrame]:
     """Find which polygons overlap without dissolving.
 
@@ -39,7 +39,7 @@ def get_polygon_clusters(
         cluster_col: Name of the resulting cluster column.
         allow_multipart: Whether to allow mutipart geometries in the gdfs.
             Defaults to False to avoid confusing results.
-        wkt_col: Whether to return the cluster column values as a string with x and y
+        as_string: Whether to return the cluster column values as a string with x and y
             coordinates. Convinient to always get unique ids.
             Defaults to False because of speed.
 
@@ -154,7 +154,7 @@ def get_polygon_clusters(
 
     concated[cluster_col] = component_mapper
 
-    if wkt_col:
+    if as_string:
         concated[cluster_col] = get_grouped_centroids(concated, groupby=cluster_col)
 
     concated = _push_geom_col(concated)
